requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.10
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.40
langchain-openai==0.3.7
langsmith==0.3.11
numpy==2.2.3
openai==1.65.2
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.0
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
wsproto==1.2.0
zstandard==0.23.0
